Fixes to get ready for upload to Seeed.
Changed IREF to 560k. (3.3V instead of 5V).

Changed the boost converter voltage divider to 16k/2.2k -- this
gives an output of 10.15V instead of 10.25V but the parts are in the
OPL.

Changed the blue LED to 0603. Didn't have blue in OPL in 0805. Kept the
others in 0805 (there's no red in 0603...).

Changed most of the resistors to 0603 (to match what was in OPL).

Add a script to generate a gerber zip in a format for Seeed.